prism-image-search/client-python/requirements.txt
setuptools~=59.6.0
grpcio

